requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.6
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.4
djangorestframework==3.13.1
dnspython==2.2.1
drf-yasg==1.20.0
eventlet==0.33.1
flower==1.0.0
greenlet==1.1.2
humanize==4.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
prometheus-client==0.14.1
prompt-toolkit==3.0.29
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
